+++ frameworks/react/components/display/calendar-event/CalendarEvent.d.ts
@@ -13,6 +13,8 @@ export interface CalendarEventProps {
 
   colorId?: CatSlot;
 
+  disabled?: boolean;
+
   actionsEnabled?: boolean;
 
   actions?: React.ReactNode;